backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
# Testing
pytest==7.4.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==19.6.2

# Code quality
flake8==6.1.0
black==23.9.1
isort==5.12.0